Extract References Utility
